Replace timeconst.bc with mktimeconst.c
Update of my decade-old patch replacing timeconst.pl with mktimeconst.c,
now removing the only user of "bc" from the build, starting from the
2015 version at:
https://github.com/landley/aboriginal/blob/master/sources/patches/linux-noperl-timeconst.patch

Not much changed: one extra iteration of the loop for nanoseconds
and different makefile plumbing calling it.

In theory this could calculate the values at runtime in a small _init
function and eliminate the header or even allow HZ to change at runtime.

Signed-off-by: Rob Landley <rob@landley.net>